+++ index.js
@@ -174,7 +174,7 @@ ObjectTemplate.setTemplateProperties = function setTemplateProperties(template, 
  * @returns {*} the object template
  */
 ObjectTemplate.create = function create(name, properties) {
-    if (typeof(name) != 'undefined' && name.name) {
+    if (name && name.name) {
         var props = name;
         name = props.name;
     }
@@ -215,7 +215,11 @@ ObjectTemplate.create = function create(name, properties) {
  * Extend and existing (parent template)
  *
  * @param {unknown} parentTemplate unknown
- * @param {unknown} name of the template
+ * @param {unknown} name the name of the template or an object with
+ *        name - the name of the class
+ *        toClient - whether the object is to be shipped to the client (with semotus)
+ *        toServer - whether the object is to be shipped to the server (with semotus)
+ *        isLocal - equivalent to setting toClient && toServer to false
  * @param {unknown} properties are the same as for create
  *
  * @returns {*} the object template
@@ -356,6 +360,18 @@ ObjectTemplate.createIfNeeded = function createTemplate (template, thisObj)
         }
         if (thisObj) {
             //var copy = new template();
+            var prototypes = [template.prototype];
+            var parent = template.__parent__;
+            while (parent) {
+                prototypes.push(parent.prototype);
+                parent = parent.__parent__;
+            }
+            for (var ix = prototypes.length - 1; ix >= 0; --ix) {
+                var props = Object.getOwnPropertyNames(prototypes[ix])
+                props.forEach(function (val, ix) {
+                    Object.defineProperty(thisObj, props[ix], Object.getOwnPropertyDescriptor(prototypes[ix], props[ix]));
+                });
+            }
             thisObj.__proto__ = template.prototype;
         }
     }